new features add, code refactored, shell completion updated

+++ imtools/SwarmClass.py
@@ -2,32 +2,55 @@ import datetime
 import os
 import yaml
 import subprocess
+import docker
 from tabulate import tabulate
 
-from imtools.IMFuncsAndConst import (CONTAINER_IOC_RUN_PATH, LOG_FILE_DIR, MOUNT_DIR, GLOBAL_SERVICE_FILE, SWARM_DIR,
-                                     CONFIG_FILE_NAME, IOC_SERVICE_FILE, PREFIX_STACK_NAME, REPOSITORY_DIR,
-                                     SWARM_BACKUP_DIR,
-                                     relative_and_absolute_path_to_abs, try_makedirs, get_manager_path, )
+from .IMConsts import (CONTAINER_IOC_RUN_PATH, LOG_FILE_DIR, MOUNT_DIR, GLOBAL_SERVICE_FILE, SWARM_DIR,
+                       IOC_SERVICE_FILE, PREFIX_STACK_NAME, REPOSITORY_DIR, SWARM_BACKUP_DIR, )
+from .IMFuncs import relative_and_absolute_path_to_abs, try_makedirs, get_manager_path
+from .ServiceDefinition import GlobalServicesList, LocalServicesList
 
 
 class SwarmManager:
     def __init__(self):
-        repository_path = os.path.join(get_manager_path(), REPOSITORY_DIR)
-        self.services = {item: SwarmService(name=item, service_type='ioc') for item in os.listdir(repository_path)}
-        self.services['log'] = SwarmService(name='log', service_type='global')
-        self.global_services = ('log',)
+        self.services = {item: SwarmService(name=item, service_type='ioc') for item in
+                         os.listdir(os.path.join(get_manager_path(), REPOSITORY_DIR))}
+        for ss in GlobalServicesList:
+            self.services[ss] = SwarmService(name=ss, service_type='global')
+        for ss in LocalServicesList:
+            name, compose_file = ss
+            self.services[ss] = SwarmService(name=name, service_type='local', compose_file=compose_file)
+        self.client = docker.from_env()
+        self.running_services = self.get_services_from_docker()
+
+    def get_services_from_docker(self):
+        services = self.client.services.list(filters={'label': f'com.docker.stack.namespace={PREFIX_STACK_NAME}'})
+        return [item for item in services]
+
+    def list_running_services(self):
+        temp_list = [item.name for item in self.running_services]
+        temp_list.sort()
+        for item in temp_list:
+            print(f'{item}', end=' ')
+        else:
+            print()
 
     def show_info(self):
-        raw_print = [["IOC", "ServiceName", "Type", "Status", ], ]
+        raw_print = [["Name", "ServiceName", "Type", "Status", ], ]
+        local_print = []
         global_print = []
         ioc_print = []
         for item in self.services.values():
             if item.service_type == 'ioc':
                 ioc_print.append([item.name, item.service_name, item.service_type, item.current_state])
-            else:
+            elif item.service_type == 'global':
                 global_print.append([item.name, item.service_name, item.service_type, item.current_state])
+            else:
+                local_print.append([item.name, item.service_name, item.service_type, item.current_state])
         ioc_print.sort(key=lambda x: x[0])
         global_print.sort(key=lambda x: x[0])
+        local_print.sort(key=lambda x: x[0])
+        raw_print.extend(local_print)
         raw_print.extend(global_print)
         raw_print.extend(ioc_print)
         print(tabulate(raw_print, headers="firstrow", tablefmt='plain'))
@@ -149,14 +172,11 @@ class SwarmManager:
                 f'. ~/.bash_aliases; '
                 f'export EPICS_IOC_LOG_FILE_NAME='
                 f'{os.path.join(CONTAINER_IOC_RUN_PATH, LOG_FILE_DIR, "$$(hostname).ioc.log")}; '
-                f'date;'
+                f'date; '
                 f'echo export EPICS_IOC_LOG_FILE_NAME=$${{EPICS_IOC_LOG_FILE_NAME}}; '
-                f'echo run iocLogServer; iocLogServer'
+                f'echo run iocLogServer; '
+                f'iocLogServer'
             ],
-            # 'environment': {
-            #     'EPICS_IOC_LOG_FILE_NAME':
-            #         f'{os.path.join(CONTAINER_IOC_RUN_PATH, LOG_FILE_DIR, f"$$(hostname).ioc.log")}',
-            # },
             'deploy': {
                 'mode': 'global',
                 'restart_policy':
@@ -191,7 +211,7 @@ class SwarmManager:
             yaml.dump(yaml_data, file, default_flow_style=False)
 
     @staticmethod
-    def get_deployed_services():
+    def get_deployed_swarm_services():
         result = subprocess.run(
             ['docker', 'service', 'ls', '-f', f'label=com.docker.stack.namespace={PREFIX_STACK_NAME}', '--format',
              '{{.Name}}'],
@@ -200,12 +220,25 @@ class SwarmManager:
         return output_list
 
     @staticmethod
+    def get_deployed_compose_services():
+        result = subprocess.run(
+            ['docker', 'compose', 'ls', ],
+            stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True)
+        output_list = result.stdout.splitlines()
+        temp_list = output_list[1:]
+        return [item.split(' ')[0] for item in temp_list]
+
+    @staticmethod
     def show_deployed_services():
         os.system(f'docker stack ps -f "desired-state=running" -f "desired-state=ready" {PREFIX_STACK_NAME}')
 
     @staticmethod
-    def show_deployed_info():
+    def show_deployed_services_detail():
         os.system(f'docker stack ps {PREFIX_STACK_NAME}')
+
+    @staticmethod
+    def show_compose_services():
+        os.system(f'docker compose ls')
 
     @staticmethod
     def show_deployed_machines():
@@ -301,41 +334,51 @@ class SwarmManager:
 
 
 class SwarmService:
-    def __init__(self, name, service_type):
+    def __init__(self, name, service_type, **kwargs):
+        """
+
+        :param name: service name.
+        :param service_type:
+            "ioc" or
+            "global"(services that should run on each node) or
+            "local"(other services that also should be run in this system)
+        """
         self.name = name
-        self.service_type = service_type
+        self.service_type = None
         self.service_name = f'{PREFIX_STACK_NAME}_srv-{name}'
         if service_type == 'ioc':
+            self.service_type = 'ioc'
             self.dir_path = os.path.join(get_manager_path(), '..', MOUNT_DIR, SWARM_DIR, self.name)
             self.service_file = IOC_SERVICE_FILE
-        else:  # 'global'
+            self.service_name = f'{PREFIX_STACK_NAME}_srv-{name}'
+        elif service_type == 'global':
             self.service_type = 'global'
             self.dir_path = os.path.join(get_manager_path(), '..', MOUNT_DIR, SWARM_DIR)
             self.service_file = GLOBAL_SERVICE_FILE
+        else:
+            self.service_type = 'local'
+            compose_file_path = kwargs.get('compose_file')
+            if not compose_file_path:
+                self.dir_path = ""
+                self.service_file = ""
+            else:
+                compose_file_path = relative_and_absolute_path_to_abs(compose_file_path)
+                self.dir_path = os.path.dirname(compose_file_path)
+                self.service_file = os.path.basename(compose_file_path)
 
-    def __repr__(self):
-        return f'SwarmService("{self.name}")'
+    def __str__(self):
+        return f'SwarmService-{self.name}(Type: {self.service_type})'
 
     @property
-    def is_available(self, verbose=False):
-        if os.path.isfile(os.path.join(self.dir_path, CONFIG_FILE_NAME)):
-            if os.path.isfile(os.path.join(self.dir_path, self.service_file)):
-                flag = True
-            else:
-                flag = False
-                if verbose:
-                    print(f'SwarmService("{self.name}").is_available: File "{self.service_file}" not exists.')
+    def is_available(self):
+        if os.path.isfile(os.path.join(self.dir_path, self.service_file)):
+            return True
         else:
-            if self.service_type == 'global' and os.path.isfile(os.path.join(self.dir_path, self.service_file)):
-                return True
-            flag = False
-            if verbose:
-                print(f'SwarmService("{self.name}").is_available: File "{CONFIG_FILE_NAME}" not exists.')
-        return flag
+            return False
 
     @property
     def is_deployed(self):
-        if self.service_name in SwarmManager.get_deployed_services():
+        if self.service_name in SwarmManager.get_deployed_swarm_services():
             return True
         else:
             return False
@@ -368,17 +411,18 @@ class SwarmService:
         else:
             print(f'SwarmService("{self.name}").deploy_service: Failed to deploy, service is not available.')
 
-    def remove(self):
+    def remove(self, remove_file=False):
         if self.is_deployed:
             print(f'SwarmService("{self.name}").remove_service: Removing this service.')
             os.system(f'docker service rm {self.service_name}')
-            if os.path.isfile(os.path.join(self.dir_path, self.service_file)):
-                try:
-                    os.remove(os.path.join(self.dir_path, self.service_file))
-                except Exception as e:
-                    print(f'SwarmService("{self.name}").remove_service: Remove swarm file failed.')
-                else:
-                    print(f'SwarmService("{self.name}").remove_service: Swarm file removed.')
+            if remove_file:
+                if os.path.isfile(os.path.join(self.dir_path, self.service_file)):
+                    try:
+                        os.remove(os.path.join(self.dir_path, self.service_file))
+                    except Exception as e:
+                        print(f'SwarmService("{self.name}").remove_service: Remove swarm file failed.')
+                    else:
+                        print(f'SwarmService("{self.name}").remove_service: Swarm file removed.')
         else:
             print(f'SwarmService("{self.name}").remove_service: Failed to remove, service is not deployed.')
 
@@ -424,5 +468,8 @@ if __name__ == '__main__':
     # s.show_ps()
     # s.show_info()
     # print(s.current_state)
-    SwarmManager().show_info()
+    # SwarmManager().show_info()
+    # print(SwarmManager().get_services_from_docker())
+    # SwarmManager().list_running_services()
     # SwarmManager.backup_swarm()
+    print(SwarmManager.get_deployed_compose_services())
